pyproject.toml
[project]
name = "localllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "jwt>=1.4.0",
    "lmcache>=0.3.9.post2",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "vllm>=0.11.0",
]

[tool.uv]
index = [
    { name = "pytorch-cu122", url = "https://download.pytorch.org/whl/cu122" }
]

[tool.uv.workspace]
members = [
    "worker",
    "api",
]

[tool.alembic]
script_location = "alembic"
